requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
djangorestframework==3.11.0
gevent==20.4.0
gevent-eventemitter==2.1
greenlet==0.4.15
idna==2.9
pip-autoremove==0.9.1
protobuf==3.11.3
psycopg2-binary==2.8.5
pycparser==2.20
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
steam==0.9.1
urllib3==1.25.9
vdf==3.3
